Fix sed incompatibilities with -i
The -i option, which edits files in place, can take a backup extension.
In the GNU implementation, if you don't pass an extension it will
replace the old file. However, the BSD implementation requires that you
pass in an extension. We can simply pass in an empty string '' to
replace the old file. However, the GNU implementation throws an error if
we pass in an empty string! Because of these terrible incompatibilities,
a temporary file is used in place of -i.