Add back old site files, convert current site
Horrendous, I know. A terrible commit, but I got annoyed. Anyways, just
trust that all the files added are pretty much the same as the old
files. The only difference is that JS files prepended with a "_" are
ignored. This is a hack so that we can keep the old files without using
them in the new site. Also, the old site has been split up into
templates, and the FAQ is no longer duplicated. Hooray! FYI, the FAQ
yaml must be in index.yaml because that's where the faq partial is
included. Also, normalize.css was renamed to a SCSS partial because
otherwise SASS will use an @import instead of just concatenating the
files together.