Change abspath to notdir and clean up filter-out
Since we cd into $(TEMPLATE_DIR), we can use notdir instead of abspath
and reduce the length of the command line output. The first argument to
the filter-out function has been replaced with $(HTML_PARTIALS) so that
there will not be an inconsistency should the naming scheme for partials
ever change (i.e. they are not prefixed with an underscore anymore).