Remove target="_blank" from Typeform link
This change was made before in e6eb116de5b382301f4d57a7e147b6e4923ff228.
Again, see https://css-tricks.com/use-target_blank/ for the rationale
behind this decision. The main reason, though, which wasn't mentioned
last time, is that the page linked to gains access to our page through
window.opener. It's not too big of an issue, but a vulnerability none
the less.